Update test README and structure
The testing directory is now separate from the main application and the
instructions for testing have been updated. It will obviously assume that shrunk
has been installed in a virtualenv but for now it won't be listed in pip.req.

Perhaps a future commit will add a setup.py; this will allow for easy testing
with:

    (virtualenv) $ python setup.py sdist
    (virtualenv) $ pip install --no-index=--find-links shrunk0.1/ shrunk

or something similar.