Update client.py
Fixed random URL generation and the way new links are created to allow for
custom URLs to be passed in. This also fixes the implementation of get_urls to
return the result properly in list format.

This adds a new custom exception; if we need more we can consider creating our
own module with custom shrunk exceptions.