Add ShrunkCursor object
In trying to abstract away the cursor interaction from the users of client.py,
we lost out on some important functionality (namely skipping results, limiting
results, and performing quick counts of the result set). The ShrunkCursor class
should be a decent compromise between the two, which allows us to interact with
the cursor more without the user knowing how it works.

The main feature this adds is easy pagination of the results.