Add original URL in link information
The link info boxes on the home page now show the original link that was
shrunken.

A future commit should change either the highlighting style or the font color as
it's difficult to read the grayed-out font when a box is hovered over.

Another future commit might want to change how the original link is displayed.