Redirect to index after link CRUD operations
Creating, modifying or deleting a link now forces a redirect to the main page.
This prevents some unintuitive behavior with refreshing after performing the
update.

Resolves #18.

This update has now removed some of the useful messages at the top of the index
that notifies the user that something has been added/changed/deleted. If we want
to add that functionality again, it will probably have to be sent in with the
GET arguments as it is no longer possible to pass in Flask variables like
before.