Add tracking for browsers, platforms and referrers to DB
We want to implement a stats view similar to goo.gl. Before, upon any
request, only the source_ip was being tracked for those accessing a URL.
Now, browsers, platforms and referrers are being tracked and added to
the MongoDB instance. Domain names are extracted with a get_domain()
function in util.py. The DB client visit() method was updated to add
this new information, and this new information was extracted from the
request objects in both appserver.py and linkserver.py.